fix(rag): use IconDefinition for Button icon and rename utils file
Pass faTrash as an IconDefinition import to the Button component
instead of a string, matching the convention used by all other Button
usages. Rename rag-environment-utils.svelte.ts to .ts since it
contains no reactive state, aligning with other pure utility files.

Fixes #1289

Made-with: Cursor
