+++ packages/renderer/src/lib/rag/rag-environment-utils.ts
